Move Follow delivery Activities into Worker
