+++ apps/worker/package.json
@@ -8,7 +8,7 @@
     "start": "node --import tsx src/index.ts",
     "test": "pnpm test:unit && pnpm test:workflow && pnpm test:database",
     "test:database": "cd ../.. && node scripts/test-db.mjs run -- pnpm --filter @kosmo/worker test:database:isolated",
-    "test:database:isolated": "cd ../.. && pnpm db:test:push && pnpm --filter @kosmo/worker exec node --import tsx --import ../../packages/core/temporal/test-client.ts --test --test-concurrency=1 src/activities.database.test.ts",
+    "test:database:isolated": "cd ../.. && pnpm db:test:push && pnpm --filter @kosmo/worker exec node --import tsx --import ../../packages/core/temporal/test-client.ts --test --test-concurrency=1 src/activities.database.test.ts src/profile-follow-activities.database.test.ts",
     "test:unit": "pnpm build && node --import tsx --test src/worker.test.ts",
     "test:workflow": "node --import tsx --test src/workflows.test.ts"
   },
@@ -17,10 +17,13 @@
     "@kosmo/fedify": "workspace:*",
     "@temporalio/worker": "^1.22.0",
     "@temporalio/workflow": "^1.22.0",
+    "drizzle-orm": "1.0.0-beta.22",
     "ts-pattern": "^5.9.0",
     "tsx": "^4.22.4"
   },
   "devDependencies": {
+    "@fedify/fedify": "^2.3.0",
+    "@fedify/vocab": "2.3.0",
     "@temporalio/client": "^1.22.0",
     "@temporalio/testing": "^1.22.0"
   }
